requirements.txt
python-telegram-bot[job-queue]>=20.0
gspread>=5.0.0
google-auth==2.23.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
oauth2client==4.1.3
python-dotenv>=0.19.0
requests==2.31.0
aiohttp==3.10.5
Flask==2.2.3
Werkzeug==2.2.3
google-api-python-client>=2.0.0